requirements.txt
rpaframework
selenium~=4.15.2
requests
openpyxl~=3.1.5
python-dotenv~=1.0.1
python-dateutil~=2.9.0.post0